+++ dockerfiles/BUILD
@@ -1,0 +1,9 @@
+docker_image(
+    name="docker",
+    source="Dockerfile.dagit",
+)
+
+docker_image(
+    name="docker0",
+    source="Dockerfile.user_pipeline",
+)
