+++ dockerfiles/Dockerfile.user_pipeline
@@ -1,0 +1,3 @@
+FROM python:3.9-slim
+
+CMD ["dagster", "api", "grpc", "-h", "0.0.0.0", "-p", "4000", ]
