+++ pyproject.toml
@@ -22,9 +22,10 @@ python = "^3.9"
 PyPika = "^0.48.9"
 dagit = "^0.14"
 dagster = "^0.14"
-dagster-airbyte = "^0.14.15"
+dagster-airbyte = "^0.14"
 dagster-aws = "^0.14"
-dagster-dbt = "^0.14.15"
+dagster-dbt = "^0.14"
+dagster-docker = "^0.14"
 dagster-gcp = "^0.14"
 dagster-ge = "^0.14"
 dagster-postgres = "^0.14"
