+++ src/ol_orchestrate/ops/elt.py
@@ -1,0 +1,11 @@
+from dagster import op
+
+
+@op
+def sync_airbyte_connection():
+    ...
+
+
+@op
+def materialize_dbt_model():
+    ...
